Tính tổng các số tự nhiên

+++ UPCODER/TINH_TONG.cpp
@@ -2,11 +2,12 @@
 using namespace std;
 
 string s;
-int ans = 0;
+long long ans = 0;
 
 int main() {
     cin >> s;
     sort(s.begin(), s.end());
+    
     do {
         ans += stoll(s);
     } while (next_permutation(s.begin(), s.end()));
